add: Added unit testing for store and reducer, tests past
