--- ReduxToolKit/src/Components/index.test.jsx
@@ -1,1 +1,0 @@
-// Template for testing
