+++ ReduxToolKit/src/Reducers/taskReducer.test.ts
@@ -1,1 +1,33 @@
-// Template for testing
+import { todoReducer } from './taskReducer';
+import { describe, it, expect } from 'vitest'; 
+import { ADD_TODO, TOGGLE_TODO, AddTodoAction, ToggleTodoAction } from './index';
+
+describe('todoReducer', () => {
+  it('should handle ADD_TODO', () => {
+    const initialState = { todos: [] };
+    const action: AddTodoAction = { type: ADD_TODO, payload: 'Test todo' };
+    const newState = todoReducer(initialState, action);
+    expect(newState.todos).toHaveLength(1);
+    expect(newState.todos[0].text).toEqual('Test todo');
+    expect(newState.todos[0].completed).toBeFalsy();
+  });
+
+  it('should handle TOGGLE_TODO', () => {
+    const initialState = {
+      todos: [
+        { id: 1, text: 'Test todo', completed: false }
+      ]
+    };
+    const action: ToggleTodoAction = { type: TOGGLE_TODO, payload: 1 };
+    const newState = todoReducer(initialState, action);
+    expect(newState.todos[0].completed).toBeTruthy();
+  });
+
+  it('should return the same state for unknown action types', () => {
+    const initialState = { todos: [] };
+    const action = { type: 'UNKNOWN_ACTION', payload: 'Test todo' } as any;
+    const newState = todoReducer(initialState, action);
+    expect(newState).toEqual(initialState);
+  });
+});
+
